requirements.txt
Flask==2.3.3
google-auth==2.23.3
google-auth-oauthlib==1.1.0
scikit-learn==1.3.0
nltk==3.8.1
gunicorn==21.2.0
beautifulsoup4==4.12.2
python-dateutil==2.8.2
numpy==1.24.3
flask-cors==4.0.0